requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
ccxt==2.7.49
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.5
pycares==4.3.0
pycparser==2.21
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
yarl==1.8.2
